tests/wpt/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.create2.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.create2.nonfinite</h1>
<p class="desc">createImageData() throws TypeError if arguments are not finite</p>


<script>
var t = async_test("createImageData() throws TypeError if arguments are not finite");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

assert_throws_js(TypeError, function() { ctx.createImageData(Infinity, 10); });
assert_throws_js(TypeError, function() { ctx.createImageData(-Infinity, 10); });
assert_throws_js(TypeError, function() { ctx.createImageData(NaN, 10); });
assert_throws_js(TypeError, function() { ctx.createImageData(10, Infinity); });
assert_throws_js(TypeError, function() { ctx.createImageData(10, -Infinity); });
assert_throws_js(TypeError, function() { ctx.createImageData(10, NaN); });
assert_throws_js(TypeError, function() { ctx.createImageData(Infinity, Infinity); });
var posinfobj = { valueOf: function() { return Infinity; } },
    neginfobj = { valueOf: function() { return -Infinity; } },
    nanobj = { valueOf: function() { return -Infinity; } };
assert_throws_js(TypeError, function() { ctx.createImageData(posinfobj, 10); });
assert_throws_js(TypeError, function() { ctx.createImageData(neginfobj, 10); });
assert_throws_js(TypeError, function() { ctx.createImageData(nanobj, 10); });
assert_throws_js(TypeError, function() { ctx.createImageData(10, posinfobj); });
assert_throws_js(TypeError, function() { ctx.createImageData(10, neginfobj); });
assert_throws_js(TypeError, function() { ctx.createImageData(10, nanobj); });
assert_throws_js(TypeError, function() { ctx.createImageData(posinfobj, posinfobj); });
t.done();

});
</script>
